removed core_utils.py, wip transformations
