+++ src/spatialdata/_dataloader/datasets.py
@@ -6,7 +6,6 @@ from torch.utils.data import Dataset
 
 from spatialdata import SpatialData
 from spatialdata._core._rasterize import rasterize
-from spatialdata._core.core_utils import get_axis_names
 from spatialdata._core.spatialdata_operations import get_transformation
 from spatialdata._core.transformations import BaseTransformation
 from spatialdata._utils import _affine_matrix_multiplication
@@ -16,6 +15,7 @@ from spatialdata.models import (
     Labels2DModel,
     Labels3DModel,
     ShapesModel,
+    get_axis_names,
     get_model,
 )
