+++ tests/_core/test_ngff_transformations.py
@@ -4,7 +4,7 @@ import json
 import numpy as np
 import pytest
 
-from spatialdata._core.core_utils import C, X, Y, Z, get_default_coordinate_system
+from spatialdata._core.ngff._utils import get_default_coordinate_system
 from spatialdata._core.ngff.ngff_coordinate_system import NgffCoordinateSystem
 from spatialdata._core.ngff.ngff_transformations import (
     NgffAffine,
@@ -18,6 +18,7 @@ from spatialdata._core.ngff.ngff_transformations import (
     NgffTranslation,
 )
 from spatialdata._types import ArrayLike
+from spatialdata.models import C, X, Y, Z
 from tests._core.conftest import (
     c_cs,
     cyx_cs,
