+++ tests/test_utils.py
@@ -14,7 +14,7 @@ from xarray import DataArray
 from spatialdata import read_zarr
 from spatialdata._core.transformations import Affine
 from spatialdata._io._utils import get_backing_files
-from spatialdata.element_utils._utils import (
+from spatialdata.element_utils.element_utils import (
     multiscale_spatial_image_from_data_tree,
     unpad_raster,
 )
